Moves HMC logic to `utils/hmc.py`
